Creating entropy function and editing test dataset
